refactor : 추상화 제거
